requirements.txt
fastapi
uvicorn
python-multipart
supabase
python-dotenv
filetype
beautifulsoup4
playwright
google_generativeai
langchain_google_genai
browser-use
python-jose
bcrypt
